Use average of scores instead of best score
previously, we'd only record the score of whichever trial had
the highest score after num_moves. This is stupid because we're
basically hoping that performing the move on the real board
will spontaneously give us that best-case scenario as well.
By recording the score we get after every trial, we can get an
average score that we can expect by performing first_move