Score from first move given extra weight
The AI used to ignore the points we could score from the first move, instead
only looking at the score that the succeeding trials can net.
This led to it making weird decisions where the first move would ignore
merging the biggest tiles because, in succeeding trials, it merges
them later on anyway.

By giving extra weight to the first move, we're preventing the bot
from just ignoring the benefits of making that first move