Fixed logging errors
- using 'continue' instead of 'pass' meant that we'd never click
resetGame after failing to click continueGame since failing
continueGame would skip right over all of resetGame
- continueGame would default to an empty string if it didn't exist,
but resetGame always had a string value. This makes it awkward to
if-else the string values. Instead, it's better to attempt clicking
the button first and THEN attempt to log stuff. If the click failed,
we'll just go into the except statement instead